Initial Build for Reduce and Completed. Fixed Error on Counters. MapReduce got updated to include new trees/functions similar to map phase

+++ CompletedCounter.java
@@ -20,7 +20,7 @@ public class CompletedCounter implements CounterInterface {
 	}
 
 	@Override
-	public void increment(String key, long n) throws RemoteException {
+	public void increment(String key, int n) throws RemoteException {
 		set.remove(key);
 		counter += n;
 	}
